Drop scaling, move clade count calc to initial_processing.r
We no longer need scaling to be done this way because we only use css-scaled counts for fitzig and the fitzig procedure handles that separately.
We still want the portion of the reads_scaling.R script that recalculates raw clade counts because it must be done following taxa filtering. This portion has been moved to initial_processing.R