Make ANCOMBC select first treatment in key as reference
Before, ANCOMBC would select the treatment to use as reference alphabetically
e.g. With treatments (control, sublethal, acute), acute would be chosen and resulting contrasts would be acute vs. control, acute vs. sublethal.
However, we want the control treatment to be the reference. This has been implemented by using the first treatment in treatment_key as reference. This introduces the invariant that the user must place the control treatment in the first position of the treatment_key, but it allows for generalization for when the control is not explicitly named 'control' (e.g. 'unexposed' in activity 2)
Also, the modules package is explicitly referenced when importing packages since the phyloseq package also has a function called 'import'