Edit p-value assignment and results output
I set the inv_treat/rep$sign$p.value to be equal to the adjusted pvalue so that the summary() function reports only results with significant adjusted p-value.
Also changed some ordering of the separate results tables so that p-value and adjusted p-value are better aligned in written output file.